CMO-99: user trace log
